requirements.txt
torch==2.6.0
transformers==4.30.2
pillow==9.5.0
flask==2.3.2
gunicorn==20.1.0